Se completo el examen

+++ app/Http/Controllers/UserController.php
@@ -1,9 +1,8 @@
 <?php
-
 namespace App\Http\Controllers;
-
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Auth;
 use App\Models\User;
 use Spatie\Permission\Models\Role;
 use DB;
@@ -31,9 +30,8 @@ class UserController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function index(Request $request) {
-        $data = User::orderBy('id', 'DESC')->paginate(10);
-        return view('users.index', compact('data'))
-                        ->with('i', ($request->input('page', 1) - 1) * 5);
+        $data = User::orderBy('id', 'DESC')->get();
+        return view('users.index', compact('data'));
     }
 
     /**
@@ -59,10 +57,8 @@ class UserController extends Controller {
             'password' => 'required|same:confirm-password',
             'roles' => 'required'
         ]);
-
         $input = $request->all();
         $input['password'] = Hash::make($input['password']);
-
         $user = User::create($input);
         $user->assignRole($request->input('roles'));
         if ($request->file('file')) {
@@ -71,7 +67,6 @@ class UserController extends Controller {
                 'url' => $url
             ]);
         }
-
         return redirect()->route('users.index')
                         ->with('success', 'User created successfully');
     }
@@ -97,7 +92,6 @@ class UserController extends Controller {
         $user = User::find($id);
         $roles = Role::pluck('name', 'name')->all();
         $userRole = $user->roles->pluck('name', 'name')->all();
-
         return view('users.edit', compact('user', 'roles', 'userRole'));
     }
 
@@ -112,8 +106,7 @@ class UserController extends Controller {
         $this->validate($request, [
             'name' => 'required',
             'email' => 'required|email|unique:users,email,' . $id,
-            'password' => 'same:confirm-password',
-            'roles' => 'required'
+            'password' => 'same:confirm-password'
         ]);
         $input = $request->all();
         if (!empty($input['password'])) {
@@ -121,7 +114,6 @@ class UserController extends Controller {
         } else {
             $input = Arr::except($input, array('password'));
         }
-
         $user = User::find($id);
         $user->update($input);
         if ($request->file('file')) {
@@ -130,13 +122,19 @@ class UserController extends Controller {
                 'url' => $url
             ]);
         }
+        if(isset($input['roles'])){
         //Se elimina el rol de usuario
         DB::table('model_has_roles')->where('model_id', $id)->delete();
         //Se asigna el rol
         $user->assignRole($request->input('roles'));
-
-        return redirect()->route('users.index')
-                        ->with('success', 'User updated successfully');
+        }
+        if(Auth::user()->hasRole('Admin')){
+            return redirect()->route('users.index')
+            ->with('success', 'User updated successfully');
+        }else{
+            return redirect()->route('users.show', $id)
+            ->with('success', 'User updated successfully');
+        }
     }
 
     /**
@@ -150,5 +148,4 @@ class UserController extends Controller {
         return redirect()->route('users.index')
                         ->with('success', 'User deleted successfully');
     }
-
 }
